Ficus rebase: Course Dashboard > Remove Answer Distribution
Removes the Course Answer Distribution (aka Problem Stats) functionality, while
leaving the Answer Distribution Reports functionality intact.

Removes utils and test helper functions into reports_manager, so that this
module continues to work.
