+++ course_dashboard/templates/course_dashboard/base_course.html
@@ -15,7 +15,6 @@
     <ul id="container-navigation" class="nav nav-pills">
         <li role="presentation" ${"class='active'" if active_tab == 'forum_activity' else '' }><a href="${reverse('course-dashboard:forum-activity', kwargs={'course_id': course_id})}">${_("Forum activity")}</a></li>
         <li role="presentation" ${"class='active'" if active_tab == 'wiki_activity' else '' }><a href="${reverse('course-dashboard:wiki-activity', kwargs={'course_id': course_id})}">${_("Wiki activity")}</a></li>
-        <li role="presentation" ${"class='active'" if active_tab == 'answers_distribution' else '' } ><a href="${reverse('course-dashboard:problem-stats:index', kwargs={'course_id': course_id})}">${_("Answers distribution")}</a></li>
         <li role="presentation" ${"class='active'" if active_tab == 'certificate_stats' else '' } ><a href="${reverse('course-dashboard:certificate-stats', kwargs={'course_id': course_id})}">${pgettext("certificate-stats","Certificates")}</a></li>
         <li role="presentation" ${"class='active'" if active_tab == 'reports' else '' } ><a href="${reverse('course-dashboard:reports-manager:dashboard', kwargs={'course_id': course_id})}">${_("Reports bucket")}</a></li>
     </ul>
