+++ course_dashboard/templates/course_dashboard/certificate-stats.html
@@ -1,7 +1,6 @@
 ## mako
 <%inherit file="base_course.html" />
 <%! from django.utils.translation import ugettext as _ %>
-<%! from course_dashboard.problem_stats.utils import percentage %>
 
 
 <%block name="extra_head">
