--- course_dashboard/templates/problem_stats/nothandled.html
@@ -1,11 +1,0 @@
-## mako
-<%! from django.utils.translation import ugettext as _ %>
-<%! from course_dashboard.problem_stats.utils import percentage %>
-
-<%inherit file="base.html" />
-
-<%block name="answers_distribution">
-    <p style='margin-bottom : 150px;'>
-        ${ _("The answers distribution for problem {} is not handled yet.").format(question_id)}
-    </p>
-</%block>
