+++ course_dashboard/urls.py
@@ -2,8 +2,6 @@ from django.conf.urls import url, patterns, include
 
 urlpatterns = patterns('course_dashboard.views',
     url(r'^forum/$', 'forum_activity', name='forum-activity'),
-    url(r'^problem_stats/', include('course_dashboard.problem_stats.urls',
-                                    namespace='problem-stats')),
     url(r'^certificate_stats/', 'certificate_stats', name='certificate-stats'),
     url(r'^reports_manager/', include('course_dashboard.reports_manager.urls',
                                       namespace='reports-manager')),
